Add posibility to mock classes
