+++ src/data_provider/cppcheckSuppress.txt
@@ -1,2 +1,2 @@
-*:*src/sysInfoLinux.cpp:298
+*:*src/sysInfoLinux.cpp:296
 *:*src/packages/pkgWrapper.h:95
