+++ src/data_provider/src/packages/iberkeleyDbWrapper.h
@@ -1,0 +1,27 @@
+/*
+ * Wazuh SYSINFO
+ * Copyright (C) 2015-2021, Wazuh Inc.
+ * March 16, 2021.
+ *
+ * This program is free software; you can redistribute it
+ * and/or modify it under the terms of the GNU General Public
+ * License (version 2) as published by the FSF - Free Software
+ * Foundation.
+ */
+
+#ifndef _IBERKELEY_DB_WRAPPER_H
+#define _IBERKELEY_DB_WRAPPER_H
+
+#include <memory>
+#include <cstring>
+#include "db.h"
+
+class IBerkeleyDbWrapper
+{
+    public:
+        virtual int32_t getRow(DBT & key, DBT & data) = 0;
+        virtual ~IBerkeleyDbWrapper();
+        IBerkeleyDbWrapper() = default;
+};
+
+#endif // _IBERKELEY_DB_WRAPPER_H
